Rozdrobení na molekuly a atomy
