+++ frontend/src/molecules/index.js
@@ -4,3 +4,4 @@ export { LoadingButton } from './LoadingButton';
 export { Quack } from './Quack';
 export { QuackForm } from './QuackForm';
 export { ReloadButton } from './ReloadButton';
+export { CommunityCards } from './CommunityCards';
