+++ frontend/src/organisms/index.js
@@ -3,3 +3,4 @@ export { SignInForm } from './SignInForm';
 export { SignUpForm } from './SignUpForm';
 export { TopNavigation } from './TopNavigation';
 export { Tickets } from './Tickets';
+export { HeadingWithButtons } from './HeadingWithButtons';
